Build the compiler runtime libraries with the NDK.
We ship these libraries in the NDK and need them to run on older
Android versions, so we have to build them against the NDK.

For libclang_rt.profile, this has never been tested on anything older
than L, so target android-21.

ASAN was supported starting with KitKat, so target android-19.

Bug: http://b/26115001
Change-Id: Ia36fb909e2e6e930014cdc49741b118375ce0011