Drop min sdk version for libprofile_rt.
If this is going to be used in the NDK it needs to target the NDK's
minimum supported version.

We should do this for libclang_rt as well, but that depends on
libunwind_llvm which can't build for ICS (missing a few things, most
importantly dl_iterate_phdr). We'll revisit this when we have the new
unwinder.

Test: external/clang/build.py
Bug: https://github.com/android-ndk/ndk/issues/294
Change-Id: I797b574e993535db7f5d0877b1f361ae9b7d539a